optional max_results parameter
The SphinxClient sets the max_result to 1000 per default. This change makes it possible to change the default setting. If you have a large resultset for a search, this is necessary, even if you only want to display the total number of results or a pagination which leads to the results > 1000